+++ grafana-plugin/src/network/oncall-api/types-generator/custom-schemas.ts
@@ -1,0 +1,3 @@
+// Custom properties not exposed by OpenAPI schema should be defined here
+
+export type CustomApiSchemas = {};