+++ grafana-plugin/src/utils/consts.ts
@@ -54,3 +54,5 @@ export enum PAGE {
 }
 
 export const TEXT_ELLIPSIS_CLASS = 'overflow-child';
+
+export const INCIDENT_HORIZONTAL_SCROLLING_STORAGE = 'isIncidentalTableHorizontalScrolling';